✨ 新增(graphql): 新增密碼和問題相關的GraphQL類型和查詢
- 在global.graphqls中加入Uint標量類型。
- 在password.graphqls中新增密碼連線、邊緣類型及建立、更新、查詢輸入類型。
- 在question.graphqls中新增問題選項類型及相關查詢。
- 在candidate.resolvers.go中新增一個取得單一候選人的解析器。
- 在password.resolvers.go中實作建立、更新和刪除密碼的解析器。
- 在question.resolvers.go中實作取得問題選項的解析器。
- 新增Go程式碼編寫指南，確保遵循社群標準和最佳實務。

+++ app/utils/password.go
@@ -35,7 +35,7 @@ var (
 // - format: 指定的密碼格式。
 // 返回值:
 // - 如果成功生成密碼則返回密碼字串切片，否則返回錯誤。
-func (p *Password) GeneratePassword(number int, length int, format string) ([]string, error) {
+func (p *Password) GeneratePassword(number uint, length uint, format string) ([]string, error) {
 	if length < 6 {
 		length = 6
 	}
@@ -59,8 +59,8 @@ func (p *Password) GeneratePassword(number int, length int, format string) ([]st
 	}
 
 	passwords := make([]string, number)
-	for i := 0; i < number; i++ {
-		password, err := generateRandomString(chars, length)
+	for i := uint(0); i < number; i++ {
+		password, err := generateRandomString(chars, int(length))
 		if err != nil {
 			return nil, err
 		}
